chore: revert refactor
